chore: Exact parseError
