+++ packages/chain-sdk/src/types/sp-xml/RequestErrorResponse.ts
@@ -1,0 +1,6 @@
+export interface RequestErrorResponse {
+  Error: {
+    Code: string;
+    Message: string;
+  };
+}
